requirements.txt
Django
psycopg2
django-markitup
markdown